core: move transfer ratio estimate to script-helpers.js
